requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-pasta==0.2.0
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.28.0
neo4j==5.28.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.41.3
streamlit==1.42.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.1
